+++ test/Test/Tasty/AutoCollect/ConfigTest.hs
@@ -9,10 +9,15 @@ module Test.Tasty.AutoCollect.ConfigTest (
   -- $AUTOCOLLECT.TEST.export$
 ) where
 
+import Control.Monad (forM_)
 import Data.Bifunctor (first)
 import Data.Char (isSpace)
 import Data.Text (Text)
 import qualified Data.Text as Text
+import qualified Data.Text.IO as Text
+import System.Directory (createDirectoryIfMissing)
+import System.FilePath (takeDirectory, (</>))
+import System.IO.Temp (withSystemTempDirectory)
 import Test.Predicates
 import Test.Predicates.HUnit
 import Test.Predicates.QuickCheck
@@ -66,16 +71,21 @@ test_prop "parseConfig strips whitespace" (ConfigPiece v) kspaces vspaces =
 
 {----- Configuration options -----}
 
+test =
+  testCase "parseConfig parses import" $
+    parseConfig "import = foo.conf, ../bar/baz.conf"
+      @?~ right (cfgImports `with` just (eq ["foo.conf", "../bar/baz.conf"]))
+
 test_prop :: ConfigPiece -> Property
 test_prop "parseConfig parses suite_name" (ConfigPiece v) =
   parseConfig ("suite_name = " <> v)
-    `satisfies` right (cfgSuiteName `with` just (eq v))
+    `satisfies` right (cfgSuiteName `with` just (just (eq v)))
 
 test_prop :: Property
 test_prop "parseConfig parses group_type" =
   forAll (elements groupTypeOptions) $ \(groupTypeName, groupType) ->
     parseConfig ("group_type = " <> groupTypeName)
-      `satisfies` right (cfgGroupType `with` eq groupType)
+      `satisfies` right (cfgGroupType `with` just (eq groupType))
 
 test_prop :: ConfigPiece -> Property
 test_prop "parseConfig errors on invalid group_type" (ConfigPiece v) =
@@ -92,24 +102,24 @@ groupTypeOptions =
 test_prop :: ConfigPiece -> Property
 test_prop "parseConfig parses strip_suffix" (ConfigPiece v) =
   parseConfig ("strip_suffix = " <> v)
-    `satisfies` right (cfgStripSuffix `with` eq v)
+    `satisfies` right (cfgStripSuffix `with` just (eq v))
 
 test_prop :: NonEmptyList HsIdentifier -> Property
 test_prop "parseConfig parses ingredients" (NonEmpty (map getHsIdentifier -> ingredients)) =
   parseConfig ("ingredients = " <> Text.intercalate "," ingredients)
-    `satisfies` right (cfgIngredients `with` eq ingredients)
+    `satisfies` right (cfgIngredients `with` just (eq ingredients))
 
 test_prop :: NonEmptyList (HsIdentifier, Spaces) -> Property
 test_prop "parseConfig strips whitespace when parsing ingredients" (NonEmpty (map (first getHsIdentifier) -> identifiers)) =
   let ingredientsVal = Text.intercalate "," . map (\(s, spaces) -> wrapSpaces spaces s) $ identifiers
       ingredients = map fst identifiers
    in parseConfig ("ingredients = " <> ingredientsVal)
-        `satisfies` right (cfgIngredients `with` eq ingredients)
+        `satisfies` right (cfgIngredients `with` just (eq ingredients))
 
 test_prop :: BoolOption -> Property
 test_prop "parseConfig parses ingredients_override (case insensitive)" option =
   parseConfig ("ingredients_override = " <> getText option)
-    `satisfies` right (cfgIngredientsOverride `with` eq (getBool option))
+    `satisfies` right (cfgIngredientsOverride `with` just (eq (getBool option)))
 
 test_prop :: ConfigPiece -> Property
 test_prop "parseConfig errors on invalid ingredients_override" (ConfigPiece v) =
@@ -128,6 +138,36 @@ test_prop "parseConfig errors on unknown keys" (ConfigPiece k) (ConfigPiece v) =
       , "ingredients_override"
       , "strip_suffix"
       ]
+
+{----- Configuration resolution -----}
+
+test =
+  testCase "resolveConfig imports config recursively" $
+    withSystemTempDirectory "tasty-autocollect-resolveConfig" $ \tmpdir -> do
+      let
+        files =
+          [
+            ( "foo/autocollect.conf"
+            ,
+              [ "import = ../base/autocollect.conf"
+              , "suite_name = foo"
+              ]
+            )
+          ,
+            ( "base/autocollect.conf"
+            ,
+              [ "suite_name = base"
+              , "ingredients = baseIngredients"
+              ]
+            )
+          ]
+      forM_ files $ \(fpRel, fileLines) -> do
+        let fp = tmpdir </> fpRel
+        createDirectoryIfMissing True (takeDirectory fp)
+        Text.writeFile fp (Text.unlines fileLines)
+      cfg <- resolveConfig (tmpdir </> "Main.hs") mempty{cfgImports = Just ["foo/autocollect.conf"]}
+      cfgSuiteName cfg @?= Just "foo"
+      cfgIngredients cfg @?= ["baseIngredients"]
 
 {----- Helpers -----}