+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Drop support for GHC 8.10
 * Added support for where clauses in `test_prop`, where the where clause may reference the generated arguments in `test_prop`
 * Don't autocollect invalid module names
+* Enable importing configuration from other files
 
 # v0.3.2.0